BIgDecimal +

+++ core/src/model/Good.java
@@ -12,7 +12,7 @@ public abstract class Good {
     private LocalDate expirationDate;
     private int quantity;
 
-    public Good(String id, String name, double unitDeliveryPrice, String category, LocalDate expirationDate, int quantity) {
+    public Good(String id, String name, BigDecimal unitDeliveryPrice, String category, LocalDate expirationDate, int quantity) {
         this.id = id;
         this.name = name;
         this.unitDeliveryPrice = unitDeliveryPrice;
@@ -29,7 +29,7 @@ public abstract class Good {
         return name;
     }
 
-    public double getUnitDeliveryPrice() {
+    public BigDecimal getUnitDeliveryPrice() {
         return unitDeliveryPrice;
     }
